# plugins/rest-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Zonvo AI REST API
  description: |
    Welcome to the Zonvo AI REST API. Everything your AI voice agents do on the phone — calling, qualifying,
    booking, promising a callback, honouring a do-not-call request — is available to your CRM, ERP or
    in-house app through `https://app.zonvo.tech/api/v1`.

    ## What you can automate

    | Goal | Endpoints |
    |------|-----------|
    | **Trigger calls** from your own app, now or at a scheduled time | `POST /calls` (with `scheduledAt`, `dynamicVariables`) |
    | **Run bulk campaigns** | `POST /campaigns`, `/campaigns/{id}/contacts`, `/start`, `/pause` |
    | **Sync leads and call outcomes to your CRM** | `GET /leads`, `PATCH /leads/{id}`, `POST /leads/{id}/notes`, `GET /calls?outcome=`, `POST /calls/{id}/outcome` |
    | **Book, confirm or cancel appointments** the agent made | `GET /appointments`, `PATCH /appointments/{id}` |
    | **Schedule callbacks** (single or in bulk, with idempotency keys) | `POST /callbacks`, `POST /callbacks/bulk`, `DELETE /callbacks?externalRef=` |
    | **Keep a do-not-call list** in sync with your compliance system | `GET /dnd`, `POST /dnd`, `POST /dnd/import`, `GET /dnd/check` |
    | **Receive real-time webhooks** for calls, appointments, callbacks and leads | `POST /webhooks`, `GET /webhooks/events` |

    ## Integrating from a CRM in 5 steps

    1. **Create an API key** — Settings → Developer → *Create API Key*. Give it only the scopes you need
...

...

...

...

...
        ]
...
servers:
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /calls:
    get:
      summary: List Calls
      description: |
        Retrieve a paginated list of all calls made from your account. Results include calls from all telephony engines (ElevenLabs, Plivo, Twilio-OpenAI).
        
        **Tip:** Use the `page` and `pageSize` parameters to navigate through large result sets efficiently.
        
        **Required Scope:** `calls:read`
      operationId: listCalls
      tags: [Calls]
      parameters:
        - name: outcome
          in: query
          description: Only calls with this outcome (see `GET /calls/outcomes`; unknown ids return 400)
          schema:
            $ref: '#/components/schemas/CallOutcome'
        - name: agentId
          in: query
          schema:
            type: string
            format: uuid
        - name: campaignId
          in: query
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          description: Calls created at or after this time (ISO 8601)
          schema:
            type: string
            format: date-time
          example: "2026-09-01T00:00:00+05:30"
        - name: to
          in: query
          description: Calls created before this time (ISO 8601)
          schema:
            type: string
            format: date-time
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: List of calls with pagination metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      summary: Trigger Call
      description: |
        Initiate an outbound call using a specified AI agent. The call will be placed immediately unless scheduled for a future time.
        
        <br />
        
        **Prerequisites:**
        - An active AI agent (flow or incoming type)
        - Sufficient credits in your account
        - An active phone number (or provide `fromNumber`)
        - The number is not on your do-not-call list (`GET /dnd/check`)

        <br />

        **Scheduling:** pass a future `scheduledAt` and the call is queued as a callback instead of being
        placed now — the response is `202` with `callbackId`; manage it through the Callbacks endpoints.

        **Personalisation:** `dynamicVariables` (≤ 20 keys, names `[A-Za-z0-9_]{1,40}`, values ≤ 300 chars) are
        substituted into the agent's prompt and first message as `{{key}}` (or `{key}`) on every engine, and listed
        to the agent as *Details for this call*. Put your own record id in `metadata.externalRef`: it is stored on
        the call, echoed in `call.completed`, and — for scheduled calls — used as the idempotency key.
        
        <br />
        
        **Telephony Engines:**
        - `elevenlabs` — Best voice quality, 29 languages
        - `plivo` — India/APAC optimized
        - `twilio-openai` — Global coverage
        
        <br />
        
        **Required Scope:** `calls:write`
      operationId: triggerCall
      tags: [Calls]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriggerCallRequest'
            examples:
              basic:
                summary: Basic call
                value:
                  agentId: "550e8400-e29b-41d4-a716-446655440000"
                  toNumber: "+14155551234"
              withEngine:
                summary: Specify engine
                value:
                  agentId: "550e8400-e29b-41d4-a716-446655440000"
                  toNumber: "+14155551234"
                  engine: "twilio-openai"
                  fromNumber: "+14155559999"
              personalised:
                summary: CRM lead with dynamic variables
                value:
                  agentId: "550e8400-e29b-41d4-a716-446655440000"
                  toNumber: "+919876543210"
                  dynamicVariables:
                    first_name: "Priya"
                    order_id: "ORD-48213"
                    delivery_date: "26 September"
                  metadata:
                    crmId: "SF-00412"
              scheduled:
                summary: Schedule for later (returns 202)
                value:
                  agentId: "550e8400-e29b-41d4-a716-446655440000"
                  toNumber: "+919876543210"
                  scheduledAt: "2026-09-29T11:00:00+05:30"
                  dynamicVariables:
                    first_name: "Priya"
      responses:
        '201':
          description: Call initiated successfully. Returns call ID and initial status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriggerCallResponse'
        '202':
          description: Call scheduled for `scheduledAt` (a callback was created)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduledCallResponse'
              example:
                success: true
                data:
                  scheduled: true
                  callbackId: "7d1a6c9e-0b2f-4e3a-9c8d-5f6a7b8c9d0e"
                  scheduledAt: "2026-09-29T05:30:00.000Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/InsufficientCredits'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Number is on the do-not-call list (`error.code` = `DO_NOT_CALL`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /calls/{id}:
    get:
      summary: Get Call Details
      description: |
        Retrieve detailed information about a specific call including:
        - Call status and duration
        - Full conversation transcript
        - AI-generated summary
        - Sentiment analysis
        - Recording URL (if available)
        
        <br />
        
        **Tip:** Poll this endpoint to track call progress. Status will transition from `queued` → `ringing` → `in-progress` → `completed`.
        
        <br />
        
        **Required Scope:** `calls:read`
      operationId: getCall
      tags: [Calls]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Complete call details with transcript and analytics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /calls/{id}/hangup:
    post:
      summary: Hangup Call
      description: |
        Terminate an active call immediately. Use this to end calls that are in progress.
        
        **Note:** This only works for calls with status `in-progress` or `ringing`. Completed or failed calls cannot be hung up.
        
        **Required Scope:** `calls:write`
      operationId: hangupCall
      tags: [Calls]
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Hangup request accepted. Call will terminate shortly.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HangupResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /calls/outcomes:
    get:
      summary: List Call Outcomes
      description: |
        The catalogue of call outcomes. `kind: agent` outcomes are set by the agent during the conversation
        (`set_call_outcome`) and may also be set through `POST /calls/{id}/outcome`; `kind: system` outcomes are
        set by the platform from telephony events (voicemail detection, no answer, transfer, booking).

        **Required Scope:** `calls:read`
      operationId: listCallOutcomes
      tags: [Calls]
      responses:
        '200':
          description: All outcome ids with labels
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallOutcomesResponse'
              example:
                success: true
                data:
                  - { id: interested, label: Interested, kind: agent }
                  - { id: not_interested, label: Not interested, kind: agent }
                  - { id: callback_requested, label: Callback requested, kind: agent }
                  - { id: wrong_number, label: Wrong number, kind: agent }
                  - { id: already_customer, label: Already a customer, kind: agent }
                  - { id: do_not_call, label: Do not call, kind: agent }
                  - { id: no_decision, label: No decision, kind: agent }
                  - { id: voicemail, label: Voicemail, kind: system }
                  - { id: no_answer, label: No answer, kind: system }
                  - { id: busy, label: Busy, kind: system }
                  - { id: failed, label: Failed, kind: system }
                  - { id: transferred, label: Transferred, kind: system }
                  - { id: appointment_booked, label: Appointment booked, kind: system }
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token (API key)
  parameters:
...
    PageParam:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSizeParam:
      name: pageSize
      in: query
      description: Items per page (max 100)
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
...
  schemas:
...
    CallOutcome:
      type: string
      description: How the call went. Agent-kind outcomes are set in conversation or via the API; system-kind by the platform.
      enum: [interested, not_interested, callback_requested, wrong_number, already_customer, do_not_call, no_decision, voicemail, no_answer, busy, failed, transferred, appointment_booked]
...